Split up dadalus code into modules
