+++ src/Constants.jl
@@ -1,0 +1,17 @@
+# DAEDALUS related constants copied over from {daedalus}
+
+module Constants
+
+export iS, iE, iIs, iIa, iH, iR, iD, N_COMPARTMENTS
+
+const iS = 1
+const iE = 2
+const iIs = 3
+const iIa = 4
+const iH = 5
+const iR = 6
+const iD = 7
+
+const N_COMPARTMENTS = 7
+
+end
