+++ src/Daedalus.jl
@@ -1,6 +1,9 @@
 module Daedalus
 
 # Write your package code here.
+include("Constants.jl")
+include("Ode.jl")
+include("Data.jl")
 include("model_daedalus.jl")
 
 end
